+++ CHANGELOG.md
@@ -4,9 +4,8 @@ Releases
 v1.20.0-dev (unreleased)
 --------------------
 
--   Added the `http.Interceptor` inbound option which allows the
-    behavior of incoming HTTP requests to be customized.
-
+-   http: Add `http.Interceptor` option to inbounds, which allows intercepting
+    incoming HTTP requests for that inbound.
 
 v1.19.2 (2017-10-10)
 --------------------